fibrotwin/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "fibrotwin"
version = "0.1.0"
description = "Minimal multiscale cardiac fibrosis simulator"
requires-python = ">=3.10"
dependencies = [
  "torch",
  "numpy",
  "matplotlib",
  "pyyaml",
]

[tool.setuptools]
packages = ["src"]

[tool.pytest.ini_options]
pythonpath = ["."]
